feat: implement welcome page
