+++ frontend/app/(beforeloggedin)/splash/_components/SessionWrapper.tsx
@@ -1,0 +1,13 @@
+"use client";
+import { SessionProvider } from "next-auth/react";
+import { ReactNode } from "react";
+import { Session } from "next-auth";
+
+type Props = {
+  session?: Session | null;
+  children: ReactNode;
+};
+
+export default function SessionWrapper({ session, children }: Props) {
+  return <SessionProvider session={session}>{children}</SessionProvider>;
+}
