+++ frontend/app/(loggedin)/_components/Footer.tsx
@@ -1,0 +1,3 @@
+export default function Footer() {
+  return <div className="flex w-dvw flex-row bg-lime-600">Footer</div>;
+}
