+++ frontend/app/(loggedin)/_components/Header.tsx
@@ -1,0 +1,3 @@
+export default function Header() {
+  return <div className="flex w-dvw flex-row bg-lime-600">Header</div>;
+}
