+++ frontend/app/(loggedin)/friend/page.tsx
@@ -1,0 +1,5 @@
+"use client";
+
+export default function Friend() {
+  return <>friend</>;
+}
