+++ frontend/app/(loggedin)/info-zip/page.tsx
@@ -1,0 +1,3 @@
+export default function InfoZip() {
+  return "info";
+}
