+++ frontend/app/(loggedin)/trade/page.tsx
@@ -1,0 +1,4 @@
+export default function Trade() {
+    return <div>Trade</div>;
+  }
+
