--- frontend/app/(root)/_components/Footer.tsx
@@ -1,3 +1,0 @@
-export default function Footer() {
-  return <div className="flex w-dvw flex-row bg-lime-600">Footer</div>;
-}
