--- frontend/app/(root)/_components/Header.tsx
@@ -1,3 +1,0 @@
-export default function Header() {
-  return <div className="flex w-dvw flex-row bg-lime-600">Header</div>;
-}
