--- frontend/app/(root)/trade/page.tsx
@@ -1,4 +1,0 @@
-export default function Trade() {
-    return <div>Trade</div>;
-  }
-
