+++ frontend/app/api/auth/[...nextauth]/route.tsx
@@ -1,13 +1,13 @@
 import NextAuth from "next-auth";
 import KakaoProvider from "next-auth/providers/kakao";
 
+const kakaoCustomProvider = KakaoProvider({
+  clientId: process.env.KAKAO_CLIENT_ID!,
+  clientSecret: process.env.KAKAO_CLIENT_SECRET!,
+});
+
 export const authOptions = {
-  providers: [
-    KakaoProvider({
-      clientId: process.env.KAKAO_CLIENT_ID!,
-      clientSecret: process.env.KAKAO_CLIENT_SECRET!,
-    }),
-  ],
+  providers: [kakaoCustomProvider],
 };
 
 const handler = NextAuth(authOptions);
