--- frontend/app/hook/useUser.tsx
@@ -1,11 +1,0 @@
-"use client";
-import { useSession } from "next-auth/react";
-
-export default function useUser() {
-  const { data, status } = useSession();
-  return {
-    userName: data?.user?.name!,
-    userImage: data?.user?.image!,
-    status: status!,
-  };
-}
